chore: Styling fixes for Article Topics
