+++ packages/article-topics/src/article-topic.js
@@ -6,11 +6,13 @@ import { withTrackEvents } from "@times-components/tracking";
 import styles from "./styles";
 
 const ArticleTopic = ({ name, onPress, slug }) => (
-  <Link onPress={e => onPress(e, { name, slug })} url={`/topic/${slug}`}>
-    <View style={styles.container}>
-      <Text style={styles.text}>{name}</Text>
-    </View>
-  </Link>
+  <View style={styles.spacer}>
+    <Link onPress={e => onPress(e, { name, slug })} url={`/topic/${slug}`}>
+      <View style={styles.container}>
+        <Text style={styles.text}>{name}</Text>
+      </View>
+    </Link>
+  </View>
 );
 
 ArticleTopic.propTypes = {
