+++ packages/article-topics/src/styles/index.android.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+import sharedStyles from "./shared";
+
+const styles = StyleSheet.create({
+  ...sharedStyles
+});
+
+export default styles;
